introduce AnimationStateListener
